@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "dev": "concurrently \"nuxt\" \"cd server && nodemon index.js\"",
+    "dev": "concurrently \"nuxt\" \"cd server && nodemon index.ts\"",
     "build": "nuxt build",
     "start": "nuxt start",
     "generate": "nuxt generate",
@@ -16,6 +16,7 @@
     "@nuxtjs/axios": "^5.13.1",
     "@nuxtjs/dotenv": "^1.4.1",
     "@prisma/client": "^3.8.1",
+    "@vue/runtime-dom": "^3.2.29",
     "body-parser": "^1.19.1",
     "bootstrap": "^4.6.1",
     "bootstrap-vue": "^2.21.2",
@@ -37,14 +38,14 @@
     "babel-eslint": "^10.1.0",
     "codeceptjs": "^3.2.3",
     "concurrently": "^7.0.0",
-    "eslint": "^8.7.0",
+    "eslint": "^8.8.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-nuxt": "^3.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^8.3.0",
-    "fibers": "^5.0.0",
+    "eslint-plugin-vue": "^8.4.0",
+    "fibers": "^5.0.1",
     "nodemon": "^2.0.15",
-    "playwright": "^1.18.0",
+    "playwright": "^1.18.1",
     "prettier": "^2.5.1",
     "prisma": "^3.8.1",
     "sass": "^1.49.0",
